Restructure util code
  ref #65
