--- src/util/result-or-fail.ts
@@ -1,11 +1,0 @@
-export declare type ResultOrFail<Error, Result> = Failure<Error> | Success<Result>
-
-export class Success<Result> {
-  err: null
-  constructor (public result: Result) {}
-}
-
-export class Failure<Error> {
-  result: null
-  constructor (public err: Error) {}
-}
